[GTK] LayerTreeHostGtk is creating a software scene graph https://bugs.webkit.org/show_bug.cgi?id=131471
Reviewed by Sergio Villar Senin.

At some point, a regression was introduced that sent the GTK+ WebKit2 scenegraph down the
software route. The rest of the file depends on the fact that the TextureMapper is the OpenGL
variety, so we should force that explicitly. This is causing crashes locally and soon we
should be able to test this directly.

* WebProcess/WebPage/gtk/LayerTreeHostGtk.cpp:
(WebKit::LayerTreeHostGtk::initialize): Force the use of the OpenGL scenegraph.

Identifier: 149553@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167072 268f45cc-cd09-0410-ab3c-d52691b4dbfc