Crash beneath DFG JIT code @ video.disney.com https://bugs.webkit.org/show_bug.cgi?id=131447
Reviewed by Geoffrey Garen.

The 32-bit path of speculateMisc() uses an 'is not int32' check followed by
'tag not less than Undefined' check.  The first check was incorrectly elided if we
knew that the value *was* an int32, when it should have been elided if we already
knew that the value *was not* an int32.

* dfg/DFGSpeculativeJIT.cpp:
(JSC::DFG::SpeculativeJIT::speculateMisc):
* tests/stress/test-spec-misc.js: Added test.
(getX):
(foo):
(bar):


Identifier: 149590@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167112 268f45cc-cd09-0410-ab3c-d52691b4dbfc