[Cocoa] Add a form delegate method corresponding to willSubmitForm https://bugs.webkit.org/show_bug.cgi?id=131718
Reviewed by Tim Horton.

* UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _setFormDelegate:]): Changed to define and create a FormClient (which derives
from API::FormClient and messages a delegate), and set it as the form client.

* UIProcess/API/Cocoa/_WKFormDelegate.h: Declared new delegate method.

* WebProcess/InjectedBundle/API/Cocoa/WKWebProcessPlugInFormDelegatePrivate.h: Added a
variant of willSubmitForm that returs a user object to pass to the delegate in the UI
process, giving it a temporary name with “new”.

* WebProcess/InjectedBundle/API/mac/WKWebProcessPlugInBrowserContextController.mm:
(-[WKWebProcessPlugInBrowserContextController _setFormDelegate:]): Changed willSubmitForm()
to check for the new bundle delegate method first and encode the user object as user data.
Moved the encoding of an NSObject <NSSecureCoding> as user data into a helper function.



Identifier: 149777@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167343 268f45cc-cd09-0410-ab3c-d52691b4dbfc