[New Multicolumn] Column sets below spanners don't repaint properly. https://bugs.webkit.org/show_bug.cgi?id=131778
Reviewed by Beth Dakin.


Source/WebCore:
Added fast/repaint/spanner-with-margins.html.

* rendering/RenderBox.cpp:
(WebCore::RenderBox::computeRectForRepaint):
Turn off some code that was running for in-flow RenderFlowThreads that made no sense.
I'm not convinced the code is right for other types of RenderFlowThreads either,
but will leave that to the named flow thread experts. :)

LayoutTests:
* fast/repaint/spanner-with-margin-expected.txt: Added.
* fast/repaint/spanner-with-margin.html: Added.



Identifier: 149869@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167439 268f45cc-cd09-0410-ab3c-d52691b4dbfc