DumpRenderTree generates black snapshots on some platforms https://bugs.webkit.org/show_bug.cgi?id=132029 <rdar://problem/16679363>
Reviewed by Simon Fraser.

* DumpRenderTree/mac/PixelDumpSupportMac.mm:
(createBitmapContextFromWebView):
DRT checks with the WebView to determine whether it can use the window's
backing store directly or (because of compositing) needs to take a WindowServer
snapshot. However, on some platforms and in some situations, the window can
be layer backed without the WebView necessarily currently being composited.
We should check with the window in addition to the WebView before going
ahead and using its backing store.



Identifier: 150108@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167691 268f45cc-cd09-0410-ab3c-d52691b4dbfc