Make slowPathAllocsBetweenGCs a runtime option. <https://webkit.org/b/132137>
Reviewed by Mark Hahnenberg.

This will make it easier to more casually run tests with this configuration
as well as to reproduce issues (instead of requiring a code mod and rebuild).
We will now take --slowPathAllocsBetweenGCs=N where N is the number of
slow path allocations before we trigger a collection.

The option defaults to 0, which is reserved to mean that we will not trigger
any collections there.

* heap/Heap.h:
* heap/MarkedAllocator.cpp:
(JSC::MarkedAllocator::doTestCollectionsIfNeeded):
(JSC::MarkedAllocator::allocateSlowCase):
* heap/MarkedAllocator.h:
* runtime/Options.h:



Identifier: 150186@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167772 268f45cc-cd09-0410-ab3c-d52691b4dbfc