Crash in platform/mac/accessibility/table-visible-rows.html https://bugs.webkit.org/show_bug.cgi?id=132146
Reviewed by Mark Lam.

Changed to use a local JSValueRef array temporary instead of a
std::make_unique<JSValueRef[]> when making an array of JSValues so that the temporary
JSValues are visited during garbage collection when the stack is scanned.  Otherwise,
the temporary values could be collected.

* DumpRenderTree/AccessibilityUIElement.cpp:
(convertElementsToObjectArray):
* WebKitTestRunner/InjectedBundle/EventSendingController.cpp:
(WTR::EventSendingController::contextClick):
* WebKitTestRunner/InjectedBundle/mac/AccessibilityUIElementMac.mm:
(WTR::convertElementsToObjectArray):


Identifier: 150232@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167819 268f45cc-cd09-0410-ab3c-d52691b4dbfc