REGRESSION (r164133): Selection doesn't paint when scrolling some pages https://bugs.webkit.org/show_bug.cgi?id=132172
Source/WebCore:
rdar://problem/16719473

Reviewed by Brent Fulgham.

Tests: fast/dynamic/remove-invisible-node-inside-selection.html
       fast/dynamic/remove-node-inside-selection.html

* editing/FrameSelection.cpp:
(WebCore::clearRenderViewSelection): Changed to take a Node& because having
this take a Position& was unnecessary and strange, when really it just needs
to take a document as an argument.
(WebCore::DragCaretController::nodeWillBeRemoved): Updated for the above.
(WebCore::FrameSelection::respondToNodeModification): Added code to set the
m_pendingSelectionUpdate flag and call RenderView::setNeedsLayout so the
selection will be recomputed after it's temporarily cleared when one of
the selected nodes is removed.

LayoutTests:

Reviewed by Brent Fulgham.

* fast/dynamic/remove-invisible-node-inside-selection-expected.html: Added.
* fast/dynamic/remove-invisible-node-inside-selection.html: Added.
* fast/dynamic/remove-node-inside-selection-expected.html: Added.
* fast/dynamic/remove-node-inside-selection.html: Added.

Identifier: 150256@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167845 268f45cc-cd09-0410-ab3c-d52691b4dbfc