<rdar://problem/16750708> REGRESSION (r159358): Crash in InjectedBundleHitTestResult::frame https://bugs.webkit.org/show_bug.cgi?id=132318
Reviewed by Andy Estes.

* UIProcess/Cocoa/NavigationState.mm:
(WebKit::NavigationState::PolicyClient::decidePolicyForNavigationAction): Fixed a typo: set
the source frame, not the destination frame.

* WebProcess/WebCoreSupport/WebFrameLoaderClient.cpp:
(WebKit::WebFrameLoaderClient::dispatchDecidePolicyForNavigationAction): Changed to get the
frame for the action not from a hit test result, because some link activation actions are
not mouse events, but rather from the event’s target.



Identifier: 150331@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167923 268f45cc-cd09-0410-ab3c-d52691b4dbfc